MM-88 refactor: 패키지 정리
